+++ src/java/ch/usi/da/paxos/storage/InMemory.java
@@ -35,37 +35,37 @@ import ch.usi.da.paxos.api.StableStorage;
  */
 public class InMemory implements StableStorage {
 
-	private int last_trimmed_instance = 0;
+	private long last_trimmed_instance = 0;
 	
-	private final Map<Integer, Decision> decided = new LinkedHashMap<Integer,Decision>(10000,0.75F,false){
+	private final Map<Long, Decision> decided = new LinkedHashMap<Long,Decision>(10000,0.75F,false){
 		private static final long serialVersionUID = -3704400228030327063L;
-			protected boolean removeEldestEntry(Map.Entry<Integer, Decision> eldest) {  
+			protected boolean removeEldestEntry(Map.Entry<Long, Decision> eldest) {  
 				return size() > 15000; // hold only 15'000 values in memory !                                 
 	}};
 	
 	@Override
-	public void put(Integer instance, Decision decision) {
+	public void put(Long instance, Decision decision) {
 		decided.put(instance, decision);
 	}
 
 	@Override
-	public Decision get(Integer instance) {
+	public Decision get(Long instance) {
 		return decided.get(instance);
 	}
 
 	@Override
-	public boolean contains(Integer instance) {
+	public boolean contains(Long instance) {
 		return decided.containsKey(instance);
 	}
 
 	@Override
-	public boolean trim(Integer instance) {
+	public boolean trim(Long instance) {
 		last_trimmed_instance = instance;
 		return true;
 	}
 
 	@Override
-	public Integer getLastTrimInstance() {
+	public Long getLastTrimInstance() {
 		return last_trimmed_instance;
 	}
